fix auto pagination test to always use three pages
The auto pagination test allowed 100 to 150 actions and tested that auto
pagination worked with 50 items per page. When 100 was randomly chosen, this
only required two page calls (+ one setup call), in all other cases three page
calls (+ one setup call). The test failed when 100 was chosen, so as a fix we
set the range to 101..150.