handle action array responses for firewall actions
Firewall action API endpoints respond with an array of actions. The current
code only expects a single action and will fallback to parsing the response as
an empty `Firewall` object. With this adjustment we also handle an array in a
field `actions` and return an array of `Action` objects to the caller.