lib: add more typed objects to server
For many resource types if an API response contains other resources we parse
that data as an object. For example, the API response for a server includes the
image info for that server. This image info can be represented as a `Image`
object.

We're also refactoring the method `EntryLoader#_load`, because it's getting too
long.