server: return `next_actions` data on create
When a server is created, it can return a field `next_actions` containing
actions that will be executed after the server was created. Most commonly, this
contains `start_server`. But for example if we directly pass networks or
firewalls, it also contains the actions for adding the server to the network or
applying the firewall to the server.

To cleanly start a server and wait for all actions to complete, we need to know
these `next_actions`.